+++ include/network_functions/PsiNorm.hpp
@@ -6,6 +6,6 @@
 namespace rbm_on_gpu {
 
 template<typename Psi_t>
-float psi_norm(const Psi_t& psi, const ExactSummation& exact_summation);
+double psi_norm(const Psi_t& psi, const ExactSummation& exact_summation);
 
 } // namespace rbm_on_gpu
