+++ include/network_functions/PsiOkVector.hpp
@@ -15,18 +15,18 @@ using namespace std;
 
 
 template<typename Psi_t>
-void psi_O_k_vector(complex<float>* result, const Psi_t& psi, const Spins& spins);
+void psi_O_k_vector(complex<double>* result, const Psi_t& psi, const Spins& spins);
 
 template<typename Psi_t, typename SpinEnsemble>
-void psi_O_k_vector(complex<float>* result, complex<float>* result_std, const Psi_t& psi, const SpinEnsemble& spin_ensemble);
+void psi_O_k_vector(complex<double>* result, complex<double>* result_std, const Psi_t& psi, const SpinEnsemble& spin_ensemble);
 
 #ifdef __PYTHONCC__
 
 template<typename Psi_t>
-inline xt::pytensor<complex<float>, 1> psi_O_k_vector_py(
+inline xt::pytensor<complex<double>, 1> psi_O_k_vector_py(
     const Psi_t& psi, const Spins& spins
 ) {
-    auto result = xt::pytensor<complex<float>, 1>(
+    auto result = xt::pytensor<complex<double>, 1>(
         std::array<long int, 1>({static_cast<long int>(psi.get_num_params())})
     );
