+++ include/network_functions/PsiVector.hpp
@@ -10,7 +10,7 @@ using namespace std;
 
 
 template<typename Psi_t>
-void psi_vector(complex<float>* result, const Psi_t& psi);
+void psi_vector(complex<double>* result, const Psi_t& psi);
 
 template<typename Psi_t>
 Array<complex_t> psi_vector(const Psi_t& psi);
