+++ python-bindings/pyRBMonGPU/main.cpp
@@ -27,7 +27,7 @@ using namespace rbm_on_gpu;
 using namespace pybind11::literals;
 
 template<unsigned int dim>
-using complex_tensor = xt::pytensor<std::complex<float>, dim>;
+using complex_tensor = xt::pytensor<std::complex<double>, dim>;
 
 // Python Module and Docstrings
 
@@ -40,7 +40,7 @@ PYBIND11_MODULE(_pyRBMonGPU, m)
             const complex_tensor<1u>&,
             const complex_tensor<1u>&,
             const complex_tensor<2u>&,
-            const float,
+            const double,
             const bool
         >())
         .def("copy", &Psi::copy)
@@ -77,7 +77,7 @@ PYBIND11_MODULE(_pyRBMonGPU, m)
         .def_readonly("hidden_spin_weight", &PsiDynamical::Link::hidden_spin_weight);
 
     py::class_<PsiDynamical>(m, "PsiDynamical")
-        .def(py::init<vector<complex<float>>, const bool>())
+        .def(py::init<vector<complex<double>>, const bool>())
         .def("copy", &PsiDynamical::copy)
         .def("add_hidden_spin", &PsiDynamical::add_hidden_spin)
         .def(
@@ -106,7 +106,7 @@ PYBIND11_MODULE(_pyRBMonGPU, m)
             const complex_tensor<1u>&,
             const vector<complex_tensor<1u>>&,
             const vector<complex_tensor<2u>>&,
-            const float,
+            const double,
             const bool
         >())
         .def("copy", &PsiDeep::copy)
